\name{house6311}
\docType{data}
\alias{house6311}
\title{ICPSR Data Set 6311: U.S. House Elections, 1896-1992}
\description{
This set contains data on the elections to the U.S. House of
Representatives, every two years between 1896 and 1992.}
\usage{house6311}
\format{A list containing 49 data frames. Each data frame has the
  following vectors: STATE (identifier), DIST (identifier),
  INC (incumbency: -1 for Rep, 1 for Dem), VOTE (fraction of vote for
  Democratic candidate), TURNOUT (total number of votes), DELSOUTH (1 if
  in the South, 0 otherwise.}
\references{ICPSR.}
\keyword{datasets}
