\name{pres.elect.15}
\docType{data}
\alias{pres.elect.15}
\alias{pres.data}
\alias{same.districts}
\alias{years}
\title{Presidential Elections by Congressional District, 1956-2012}
\description{
This set contains data on the elections for the President of the United
States, broken down by congressional district, for each election between
1956 and 2012.}
\usage{pres.elect.15}
\format{
  pres.data contains a list of 15 data frames corresponding to each
  election.

  years is a list of years for each election, also the names of the data
  frames in pres.data.

  same.districts is a vector where 1 indicates that this election uses
  the same congressional district map as the election that preceded it.

}
\references{Thomas, Gelman, King, Katz (2012) ``Estimating Partisan Bias
  of the Electoral College Under Proposed Changes in Elector
  Apportionment'', Statistics, Politics and Policy.}
\keyword{datasets}
